switch to only having one Java project
Multiple projects (working inside of the non-Spring project) was causing workspace compilation problems with vscode.
