--- .devcontainer/java-spring-boot/devcontainer.json
@@ -1,29 +1,0 @@
-{
-	"name": "Java Spring Boot",
-	"image": "mcr.microsoft.com/devcontainers/universal:2-focal",
-	"containerEnv": {
-      "DISABLE_JAVA_BUILD": "true"
-	},
-	"postCreateCommand": "cd java-spring-boot && make init",
-	"customizations": {
-		"vscode": {
-			"settings": {
-				"java.server.launchMode": "Standard",
-				"liveshare.allowGuestDebugControl": true,
-				"liveshare.allowGuestTaskControl": true,
-				"liveshare.languages.allowGuestCommandControl": true,
-				"liveshare.publishWorkspaceInfo": true,
-				"extensions.ignoreRecommendations": true,
-				"workbench.startupEditor": "readme",
-				"redhat.telemetry.enabled": false
-			},
-			"extensions": [
-				"ms-vsliveshare.vsliveshare",
-				"vscjava.vscode-java-pack",
-				"vscjava.vscode-java-test",
-				"vscjava.vscode-maven",
-				"ms-azuretools.vscode-docker"
-			]
-		}
-	}
-}
