--- java-spring-boot/Makefile
@@ -1,10 +1,0 @@
-init:
-	gradle build
-
-verify:
-	gradle test
-
-run:
-	gradle bootRun
-
-.PHONY: init verify run
