--- java-spring-boot/README.md
@@ -1,18 +1,0 @@
-# Java Spring Boot Project
-
-## Tech Stack
-- Java 21
-- JUnit 5
-- Spring Boot 3.2 
-    - Web
-    - JPA
-    - Lombok
-- Gradle
-
-## Common Commands
-- `make init`
-    - build the project
-- `make verify`
-    - run tests
-- `make run`
-    - start the server
