+++ java/README.md
@@ -3,10 +3,16 @@
 ## Tech Stack
 - Java 21
 - JUnit 5
-- Maven
+- Spring Boot 3.2 
+    - Web
+    - JPA
+    - Lombok
+- Gradle
 
 ## Common Commands
 - `make init`
     - build the project
 - `make verify`
     - run tests
+- `make run`
+    - start the server
