+++ java/src/main/java/com/sourceallies/interview/Solution.java
@@ -1,12 +1,13 @@
 package com.sourceallies.interview;
 
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
 public class Solution {
-
-    public static void main(String[] args) {
-        System.out.println("Hello world");
-    }
-
-    public String greeting(String name) {
-        return "Hello " + name;
+    
+    @GetMapping
+    public String getGreeting() {
+        return "Hello, world!";
     }
 }
